requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==4.4.2
distro==1.9.0
docopt==0.6.2
edge-tts==6.1.10
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
moviepy==1.0.3
multidict==6.0.5
numpy==1.26.4
openai==1.13.3
opencv-python==4.9.0.80
pillow==10.2.0
proglog==0.1.10
pydantic==2.6.3
pydantic_core==2.16.3
python-dotenv==1.0.1
requests==2.31.0
setuptools==69.1.1
sniffio==1.3.1
termcolor==2.4.0
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
webvtt-py==0.4.6
yarl==1.9.4
